requirements.txt
streamlit
pdfminer.six
python-docx
sentence-transformers
requests
reportlab
